CO-3 add transformators for conversion between metric and imperial system

+++ src/main/java/eu/borostomi/mongodbdemo/dto/IngredientDto.java
@@ -4,12 +4,14 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @ToString
 public class IngredientDto {
     private String name;
-    private Integer amount;
+    private BigDecimal amount;
     private String unit;
     private Boolean isReplaceable;
 }
